Create README.md
A Sudoku solver is a specialized computer program designed to autonomously navigate and solve Sudoku puzzles, the popular number-placement game. These puzzles consist of a 9x9 grid subdivided into nine 3x3 subgrids, and the challenge lies in filling in the grid with digits from 1 to 9 so that each row, column, and 3x3 subgrid contains all the numbers without repetition. Leveraging logical algorithms and strategies, the solver meticulously analyzes the puzzle, eliminates possibilities through deduction, and strategically makes educated guesses when required. The solver's objective is to unveil a unique solution to the puzzle, showcasing the application of computational problem-solving techniques to unravel the complexities of this classic and intellectually stimulating game.